Refactor the three main layouts for easier comprehension. Also make some SEO optimisations.
