--- app/helpers/public_helper.rb
@@ -1,2 +1,0 @@
-module PublicHelper
-end
