+++ app/views/app/posts/_post.html.erb
@@ -6,7 +6,7 @@
     <div>
       <%= link_to edit_app_post_path(post, page: params[:page]), title: "Click to edit this post" do %>
         <div class="font-medium pe-2 hover:underline">
-          <%= post_title(post) %>
+          <%= post.display_title %>
         </div>
       <% end %>
       <%= render "tags", post: post %>
