--- app/views/layouts/_head.html.erb
@@ -1,63 +1,0 @@
-<meta charset="utf-8">
-<title>
-<%= page_title %>
-</title>
-<meta name="viewport" content="width=device-width,initial-scale=1">
-
-<meta name="description" content="<%= meta_description %>">
-<meta name="keywords" content="open-source blog, email-first publishing, no-code blogging, privacy-first platform, exportable blog, minimalist blogging">
-
-<% if @blog&.google_site_verification.present? %>
-<meta name="google-site-verification" content="<%= @blog.google_site_verification %>" />
-<% end %>
-
-<!-- Open Graph / Facebook -->
-<meta property="og:type" content="<%= page_type %>">
-<meta property="og:url" content="<%= request.original_url %>">
-<meta property="og:title" content="<%= page_title %>">
-<meta property="og:description" content="<%= meta_description %>">
-
-<meta property="twitter:url" content="<%= request.original_url %>">
-<meta property="twitter:title" content="<%= page_title %>">
-<meta property="twitter:description" content="<%= meta_description %>">
-
-<% if has_open_graph_image? %>
-  <meta property="og:image" content="<%= open_graph_image %>">
-  <meta property="twitter:image" content="<%= open_graph_image %>">
-  <meta property="twitter:card" content="summary_large_image">
-<% else %>
-  <meta property="twitter:card" content="summary">
-<% end %>
-
-<%= render "layouts/structured_data" %>
-
-<%= yield :fediverse %>
-
-<link rel="canonical" href="<%= canonical_url %>" />
-
-<%= render "layouts/rss" %>
-
-<%= render "layouts/favicons" %>
-<%= csrf_meta_tags %>
-<%= csp_meta_tag %>
-
-<%= font_preload_links %>
-<%= stylesheet_link_tag "tailwind", font_stylesheet, "data-turbo-track": "reload" %>
-
-<%= stylesheet_link_tag "application", "data-turbo-track": "reload" %>
-<%= javascript_importmap_tags %>
-
-<% if @blog.present? %>
-  <% unless @blog.user.search_indexable? %>
-    <meta name="robots" content="noindex, nofollow">
-  <% else %>
-    <%= yield :robots_meta %>
-  <% end %>
-
-  <% if @blog.subdomain == "olly" %>
-  <script defer data-domain="olly.pagecord.com" src="https://plausible.io/js/script.js"></script>
-  <% end %>
-  <% if @blog.subdomain == "blog" %>
-  <script defer data-domain="blog.pagecord.com" src="https://plausible.io/js/script.js"></script>
-  <% end %>
-<% end %>
