+++ app/views/layouts/app.html.erb
@@ -2,18 +2,21 @@
 <html lang="en">
   <head>
     <meta charset="utf-8">
-    <title>Pagecord :: Admin</title>
     <meta name="viewport" content="width=device-width,initial-scale=1">
-    <%= render "layouts/default_favicons" %>
-    <%= render "layouts/manifest" %>
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
+
+    <title>Pagecord :: Admin</title>
+
+    <%= render "layouts/default_favicons" %>
+    <%= render "layouts/manifest" %>
+
     <%= stylesheet_link_tag "tailwind", "inter", "data-turbo-track": "reload" %>
     <%= stylesheet_link_tag "application", "data-turbo-track": "reload" %>
     <link href="https://cdn.jsdelivr.net/npm/@yaireo/tagify/dist/tagify.css" rel="stylesheet" type="text/css"/>
     <%= javascript_importmap_tags %>
-    <script src="https://cdn.paddle.com/paddle/v2/paddle.js"></script>
 
+    <script src="https://cdn.paddle.com/paddle/v2/paddle.js"></script>
     <%= yield :billing %>
   </head>
