+++ app/views/layouts/sessions.html.erb
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8">
-    <title>Pagecord</title>
+    <title><%= content_for?(:title) ? content_for(:title) : "Pagecord" %></title>
     <meta name="viewport" content="width=device-width,initial-scale=1">
     <%= render "layouts/default_favicons" %>
     <%= render "layouts/manifest" %>
