+++ app/views/public/pagecord_vs_substack.html.erb
@@ -6,10 +6,6 @@
   Looking for a Substack alternative? Pagecord lets you blog by email with zero setup. Compare features vs Substack and see why bloggers are switching.
 <% end %>
 
-<%= content_for :meta_keywords do %>
-  Substack alternative, blog by email, simple blogging platform, Pagecord vs Substack
-<% end %>
-
 <div class="prose dark:prose-invert mb-4 sm:mb-8 px-4">
   <h1 class="text-center">Pagecord: A Simple Substack Alternative</h1>
