+++ app/views/public/pagecord_vs_wordpress.html.erb
@@ -6,10 +6,6 @@
   Looking for a WordPress alternative? Pagecord lets you blog by email with zero setup. Compare features vs WordPress and see why bloggers are switching.
 <% end %>
 
-<%= content_for :meta_keywords do %>
-  WordPress alternative, blog by email, simple blogging platform, Pagecord vs WordPress
-<% end %>
-
 <div class="prose dark:prose-invert mb-4 sm:mb-8 px-4">
   <h1 class="text-center">Pagecord: A Simple WordPress Alternative</h1>
